Streamlit_app/requirements.txt
streamlit
pandas
geopandas
folium
plotly
folium
streamlit_folium
json5
Pillow
